lbd-xoltex-generar-py-llaves/requirements.txt
pycryptodome
python-jose
requests
pydantic
python-dotenv
fastapi
uvicorn
boto3